+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Security` in case of vulnerabilities.
 
 ## [Unreleased]
+### Changed
+- move "central" download functions from downloader to file_directory_functions
 
 ## [0.8.1] - 2021-09-10
 ### Fixed